pyproject.toml
[project]
name = "discord-gemini-bot"
version = "0.1.0"
description = "A Discord bot powered by Google Gemini 3 and Veo"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord-py>=2.6.4",
    "google-genai>=1.52.0",
    "pillow>=12.0.0",
    "python-dotenv>=1.2.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pyright>=1.1.407",
]

[project.scripts]
start = "src.main:main"
